GCI-984: Reduce duplication in listeners.
